+++ native/src/main/java/io/ballerina/stdlib/graphql/runtime/engine/EngineUtils.java
@@ -27,6 +27,7 @@ import io.ballerina.runtime.api.types.Field;
 import io.ballerina.runtime.api.types.IntersectionType;
 import io.ballerina.runtime.api.types.RecordType;
 import io.ballerina.runtime.api.types.ResourceMethodType;
+import io.ballerina.runtime.api.types.ServiceType;
 import io.ballerina.runtime.api.types.Type;
 import io.ballerina.runtime.api.types.UnionType;
 import io.ballerina.runtime.api.utils.StringUtils;
@@ -38,12 +39,18 @@ import io.ballerina.runtime.api.values.BString;
 import io.ballerina.runtime.api.values.BValue;
 import io.ballerina.stdlib.graphql.commons.types.Schema;
 import io.ballerina.stdlib.graphql.commons.utils.SdlSchemaStringGenerator;
+import io.ballerina.stdlib.graphql.runtime.engine.meta.Resource;
+import io.ballerina.stdlib.graphql.runtime.engine.meta.ServiceAnalyzer;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 
 import static io.ballerina.runtime.api.TypeTags.SERVICE_TAG;
+import static io.ballerina.stdlib.graphql.runtime.engine.Engine.RESOURCE_MAP;
 import static io.ballerina.stdlib.graphql.runtime.engine.Engine.getDecodedSchema;
+import static io.ballerina.stdlib.graphql.runtime.utils.Utils.ERROR_TYPE;
+import static io.ballerina.stdlib.graphql.runtime.utils.Utils.createError;
 
 /**
  * This class provides utility functions for Ballerina GraphQL engine.
@@ -114,6 +121,15 @@ public class EngineUtils {
     // Root operation types
     public static final String OPERATION_QUERY = "query";
     public static final String OPERATION_SUBSCRIPTION = "subscription";
+
+    public static void analyzeServices(BArray services) {
+        for (int i = 0; i < services.size(); i++) {
+            BObject service = (BObject) services.get(i);
+            ServiceAnalyzer serviceAnalyzer = new ServiceAnalyzer((ServiceType) TypeUtils.getType(service));
+            serviceAnalyzer.analyze();
+            service.addNativeData(RESOURCE_MAP, serviceAnalyzer.getResourceMap());
+        }
+    }
 
     static boolean isPathsMatching(ResourceMethodType resourceMethod, List<String> paths) {
         String[] resourcePath = resourceMethod.getResourcePath();
@@ -255,5 +271,16 @@ public class EngineUtils {
 
     public static Object getResult(BObject executorVisitor) {
         return executorVisitor.get(RESULT_FIELD);
+    }
+
+    public static Object getFieldComplexity(BObject engine, BString coordinate) {
+        BObject serviceObject = (BObject) engine.getNativeData(GRAPHQL_SERVICE_OBJECT);
+        Map<String, Resource> resourceMap = (Map<String, Resource>) serviceObject.getNativeData(RESOURCE_MAP);
+        String coordinateKey = coordinate.getValue();
+        if (resourceMap.containsKey(coordinateKey)) {
+            Resource resource = resourceMap.get(coordinateKey);
+            return resource.complexity();
+        }
+        return createError("Complexity not found for the resource: " + coordinateKey, ERROR_TYPE);
     }
 }